Make top-level statements valid entry points
This is a complementary patch in order to update some
tests for an etsfrontend issue.

The etsfrontend issue is about putting
a main() function into etsglobal in case it is missing
and there are top level statemernts that we wish to
run as an entrypoint.

Fixes #15343 internal issue.

Issue: https://gitee.com/openharmony/arkcompiler_ets_frontend/issues/I97ZGD

Testing:
- created some new runtime tests to test new cases
and the order of running: test/runtime/ets/entrypoint*.ets
- some expected files of parser tests without main
functions have been updated
- one cts-test param has been updated to avoid clashing
with an interface named `main`
- an interop expect file is changed.

Signed-off-by: Csaba Hurton <csaba.hurton@huawei.com>

Change-Id: Ibbfd4362b6b57992e2a59fcd6ecc192d3b0c9cf5