Pre-pass to help PEA deal with the reference phi
PEA cannot do it's work if a LoadObject instruction has
an object type phi as it's input and there is SaveState
between the said phi and the NewObject it is referencing.
Create loads in each of the phi input blocks that reference
the original objects and "load-phis" referencing these
loads. There will be no SaveState instructions between
phi and load instructions.

Issue: IA6H0W

Signed-off-by: Aaron Simkin <aaron.simkin@huawei-partners.com>