Correct CTS test of interface properties
Issue: https://gitee.com/openharmony/arkcompiler_ets_frontend/issues/IBT8C4

Reason:
1. Contradict with spec 9.1.5:  Simple_form.sts  Readonly_simple_form_0.sts
A class must implement all properties from all super interfaces that are always defined as a getter, a setter, or both.
2. Contradict with spec 9.6 and 10.5 interface properties (the form of getter) : getter_setter_form.sts
A compile-time error occurs if: The body of a class declaration declares a method but the name of that method is already used for a field in the same declaration.

Description:
Mark the tests as negative and remove them from ignored list.

Signed-off-by: wangyuxin <wangyuxin53@huawei.com>