Split the flags and thread status management
If we use the split scheme, the CAS in StoreStatus()
is unneded, so we can make it faster, just by using regular
"store".

Also introduced asStructNonvolatile, to avoid
unnecessary "volatile" acceses, for example if
we access "status/flags" as members of local variable.

Open question is "packed" structures, as atomic_uint16_t
isn't a POD type, "packed" cannot be used, but it looks like
we don't need this, as the union is 4-bytes aligned, and
each atomic_uint16_t should be 2-bytes aligned, as it's basically just
"uin16_t" under the hood, so no padding needed.

In terms of C++, this part might be done in a more strict
formal way, using "alignas(...) " for instance.

Issue: https://gitee.com/openharmony/arkcompiler_runtime_core/issues/IBVE2R
Change-Id: I71dcb6f275fce63f1b7e32d908e1308345ae3091
Signed-off-by: Anatoly Stepanov <stepanov.anatoly@huawei.com>