[LLVM AOT] Improved MemBarrier pass
Issue: https://gitee.com/openharmony/arkcompiler_runtime_core/issues/ICGH19

DMB's at the end of basic blocks may be sinked to their successors if at least one successor path is protected. We call that a path is protected if all data allocated at the end of a basic block would be protected by a DMB in the corresponding successor before any access to it.

This implementation only concerns about basic blocks when the terminator is conditional branch with 2 successors and there are memory allocations after the last access of protected data (in which the former MemBarrier pass would place an DMB after the last memory allocation). 2 possible optimizations introduced are:

1. If the 2 successors paths are both protected, then we can safely delete the last DMB.

2. If one successor path is protected, and the other is not protected but only have 1 predecessor, then we may sink the last DMB in the predecessor to the beginning of the unprotected successor.

Signed-off-by: HanyangLu <hanyang.lu@huawei.com>